Web Scraping from eBay
Once upon a time, there was a curious internet user who wanted to learn how eBay's search engine worked. They opened their browser's development tools and analyzed the network traffic when searching for "amazon gift card" on eBay. They discovered that eBay used a GET request to search and identified the URL variable that represented the search term.

They also found a way to display sold items only by selecting "Show only" on the left-hand side of the page and discovered the URL variable that represented this option. They even came up with a shorter URL that produced the same search results and observed how the URL changed when clicking on the next search result page.

The user was intrigued and wanted to explore this further. They decided to write a program that would automatically download the first 10 pages of search results for sold "amazon gift card" and save each page to a separate file. They added a 10 second pause between requests to mimic human behavior and avoid getting blocked by eBay.

Next, the user wrote code to parse each downloaded page using a Python or Java soup-object. They used the soup-object and their knowledge of the search results page to identify the title, price, and shipping price of each item.

But the user wasn't satisfied yet. They wanted to know if any gift cards sold above face value. Using RegEx, they extracted the value of a gift card from its title and compared it to its price + shipping. If the value was less than the price + shipping, then the gift card sold above face value. The user printed this information to the screen and calculated the fraction of Amazon gift cards that sold above face value.

In the end, the user had a better understanding of how eBay's search engine worked and was able to automate the process of searching for sold "amazon gift card" and identifying gift cards that sold above face value. They also learned that a significant fraction of gift cards sold above face value, which could be explained by factors such as demand, scarcity, and resale value.